Switch to PSR-4, create a stub of a Application object, add shell.
