+++ platform
@@ -7,34 +7,8 @@ define('CLI_ROOT', __DIR__);
 
 require 'vendor/autoload.php';
 
-use CommerceGuys\Command\InitCommand;
-use CommerceGuys\Command\EnvironmentBackupCommand;
-use CommerceGuys\Command\EnvironmentBranchCommand;
-use CommerceGuys\Command\EnvironmentDeleteCommand;
-use CommerceGuys\Command\EnvironmentListCommand;
-use CommerceGuys\Command\EnvironmentMergeCommand;
-use CommerceGuys\Command\EnvironmentSynchronizeCommand;
-use CommerceGuys\Command\GetCommand;
-use CommerceGuys\Command\ProjectDeleteCommand;
-use CommerceGuys\Command\ProjectListCommand;
-use CommerceGuys\Command\SshKeyAddCommand;
-use CommerceGuys\Command\SshKeyDeleteCommand;
-use CommerceGuys\Command\SshKeyListCommand;
-use Symfony\Component\Console\Application;
+use CommerceGuys\Platform\Cli\Application;
+
 
 $application = new Application();
-$application->add(new InitCommand);
-$application->add(new GetCommand);
-$application->add(new EnvironmentBackupCommand);
-$application->add(new EnvironmentBranchCommand);
-$application->add(new EnvironmentDeleteCommand);
-$application->add(new EnvironmentListCommand);
-$application->add(new EnvironmentMergeCommand);
-$application->add(new EnvironmentSynchronizeCommand);
-$application->add(new ProjectDeleteCommand);
-$application->add(new ProjectListCommand);
-$application->add(new SshKeyAddCommand);
-$application->add(new SshKeyDeleteCommand);
-$application->add(new SshKeyListCommand);
 $application->run();
-
