+++ src/Command/EnvironmentDeleteCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CommerceGuys\Command;
+namespace CommerceGuys\Platform\Cli\Command;
 
 use Guzzle\Http\ClientInterface;
 use Symfony\Component\Console\Input\InputArgument;
@@ -8,14 +8,14 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Yaml\Dumper;
 
-class EnvironmentSynchronizeCommand extends EnvironmentCommand
+class EnvironmentDeleteCommand extends EnvironmentCommand
 {
 
     protected function configure()
     {
         $this
-            ->setName('environment:synchronize')
-            ->setDescription('Synchronize an environment.')
+            ->setName('environment:delete')
+            ->setDescription('Delete an environment.')
             ->addArgument(
                 'project-id',
                 InputArgument::OPTIONAL,
@@ -38,26 +38,14 @@ class EnvironmentSynchronizeCommand extends EnvironmentCommand
             return;
         }
 
-        $dialog = $this->getHelperSet()->get('dialog');
-        $syncCodeText = "Synchronize code? [Y/N] ";
-        $syncDataText = "Synchronize data? [Y/N] ";
-        $syncCode = $dialog->askConfirmation($output, $syncCodeText, false);
-        $syncData = $dialog->askConfirmation($output, $syncDataText, false);
-        if (!$syncCode && !$syncData) {
-            $output->writeln("<error>You must synchronize at least code or data.</error>");
-            return;
-        }
-
-        $params = array(
-            'synchronize_code' => $syncCode,
-            'synchronize_data' => $syncData,
-        );
         $client = $this->getPlatformClient($this->environment['endpoint']);
-        $client->synchronizeEnvironment($params);
+        $client->deleteEnvironment();
+        // Refresh the stored environments, to trigger a drush alias rebuild.
+        $this->getEnvironments($this->project, TRUE);
 
         $environmentId = $input->getArgument('environment-id');
         $message = '<info>';
-        $message = "\nThe environment $environmentId has been synchronized. \n";
+        $message = "\nThe environment $environmentId has been deleted. \n";
         $message .= "</info>";
         $output->writeln($message);
     }
