+++ src/Command/EnvironmentListCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CommerceGuys\Command;
+namespace CommerceGuys\Platform\Cli\Command;
 
 use Guzzle\Http\ClientInterface;
 use Symfony\Component\Console\Input\InputArgument;
@@ -8,24 +8,20 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Yaml\Dumper;
 
-class EnvironmentBackupCommand extends EnvironmentCommand
+class EnvironmentListCommand extends EnvironmentCommand
 {
 
     protected function configure()
     {
         $this
-            ->setName('environment:backup')
-            ->setDescription('Backup an environment.')
+            ->setName('environment:list')
+            ->setDescription('Get a list of all environments.')
             ->addArgument(
                 'project-id',
                 InputArgument::OPTIONAL,
                 'The project id'
-            )
-            ->addArgument(
-                'environment-id',
-                InputArgument::OPTIONAL,
-                'The environment id'
             );
+        ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -38,13 +34,19 @@ class EnvironmentBackupCommand extends EnvironmentCommand
             return;
         }
 
-        $client = $this->getPlatformClient($this->environment['endpoint']);
-        $client->backupEnvironment();
+        $rows = array();
+        foreach ($this->getEnvironments($this->project, TRUE) as $environment) {
+            $row = array();
+            $row[] = $environment['id'];
+            $row[] = $environment['title'];
+            $row[] = $environment['_links']['public-url']['href'];
+            $rows[] = $row;
+        }
 
-        $environmentId = $input->getArgument('environment-id');
-        $message = '<info>';
-        $message = "\nA backup of environment $environmentId has been created. \n";
-        $message .= "</info>";
-        $output->writeln($message);
+        $table = $this->getHelperSet()->get('table');
+        $table
+            ->setHeaders(array('ID', 'Name', "URL"))
+            ->setRows($rows);
+        $table->render($output);
     }
 }
