+++ src/Command/EnvironmentMergeCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CommerceGuys\Command;
+namespace CommerceGuys\Platform\Cli\Command;
 
 use Guzzle\Http\ClientInterface;
 use Symfony\Component\Console\Input\InputArgument;
@@ -8,14 +8,14 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Yaml\Dumper;
 
-class EnvironmentDeleteCommand extends EnvironmentCommand
+class EnvironmentMergeCommand extends EnvironmentCommand
 {
 
     protected function configure()
     {
         $this
-            ->setName('environment:delete')
-            ->setDescription('Delete an environment.')
+            ->setName('environment:merge')
+            ->setDescription('Merge an environment.')
             ->addArgument(
                 'project-id',
                 InputArgument::OPTIONAL,
@@ -39,13 +39,13 @@ class EnvironmentDeleteCommand extends EnvironmentCommand
         }
 
         $client = $this->getPlatformClient($this->environment['endpoint']);
-        $client->deleteEnvironment();
+        $client->mergeEnvironment();
         // Refresh the stored environments, to trigger a drush alias rebuild.
         $this->getEnvironments($this->project, TRUE);
 
         $environmentId = $input->getArgument('environment-id');
         $message = '<info>';
-        $message = "\nThe environment $environmentId has been deleted. \n";
+        $message = "\nThe environment $environmentId has been merged. \n";
         $message .= "</info>";
         $output->writeln($message);
     }
