+++ src/Command/EnvironmentSynchronizeCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CommerceGuys\Command;
+namespace CommerceGuys\Platform\Cli\Command;
 
 use Guzzle\Http\ClientInterface;
 use Symfony\Component\Console\Input\InputArgument;
@@ -8,14 +8,14 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Yaml\Dumper;
 
-class EnvironmentMergeCommand extends EnvironmentCommand
+class EnvironmentSynchronizeCommand extends EnvironmentCommand
 {
 
     protected function configure()
     {
         $this
-            ->setName('environment:merge')
-            ->setDescription('Merge an environment.')
+            ->setName('environment:synchronize')
+            ->setDescription('Synchronize an environment.')
             ->addArgument(
                 'project-id',
                 InputArgument::OPTIONAL,
@@ -38,14 +38,26 @@ class EnvironmentMergeCommand extends EnvironmentCommand
             return;
         }
 
+        $dialog = $this->getHelperSet()->get('dialog');
+        $syncCodeText = "Synchronize code? [Y/N] ";
+        $syncDataText = "Synchronize data? [Y/N] ";
+        $syncCode = $dialog->askConfirmation($output, $syncCodeText, false);
+        $syncData = $dialog->askConfirmation($output, $syncDataText, false);
+        if (!$syncCode && !$syncData) {
+            $output->writeln("<error>You must synchronize at least code or data.</error>");
+            return;
+        }
+
+        $params = array(
+            'synchronize_code' => $syncCode,
+            'synchronize_data' => $syncData,
+        );
         $client = $this->getPlatformClient($this->environment['endpoint']);
-        $client->mergeEnvironment();
-        // Refresh the stored environments, to trigger a drush alias rebuild.
-        $this->getEnvironments($this->project, TRUE);
+        $client->synchronizeEnvironment($params);
 
         $environmentId = $input->getArgument('environment-id');
         $message = '<info>';
-        $message = "\nThe environment $environmentId has been merged. \n";
+        $message = "\nThe environment $environmentId has been synchronized. \n";
         $message .= "</info>";
         $output->writeln($message);
     }
